requirements.txt
async-timeout==5.0.1
bcrypt==4.1.2
black==24.8.0
blinker==1.8.2
boto3==1.34.34
botocore==1.34.34
cachelib==0.13.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
debugpy==1.8.13
dnspython==2.6.1
et-xmlfile==2.0.0
flask==3.0.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
flask-redis==0.4.0
flask-session==0.8.0
gunicorn==23.0.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==2.1.5
msgspec==0.18.6
mypy-extensions==1.1.0
numpy==1.24.4
openpyxl==3.1.2
packaging==24.2
pandas==2.0.3
passlib==1.7.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
psutil==7.0.0
pycparser==2.22
pymongo==4.10.1
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
qrcode==7.4.2
redis==5.2.1
requests==2.32.3
s3transfer==0.10.4
scrypt==0.8.27
six==1.17.0
tomli==2.2.1
typing-extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
werkzeug==3.0.1
zipp==3.20.2
